Split java.util.Formatter into Impl class to handle problem with inner enum defintion
